Fix for uninstalling core library files (custom document formats and extended Fabrik database driver)

+++ administrator/components/com_fabrik/com_fabrik.manifest.class.php
@@ -4,7 +4,8 @@
  *
  * @package     Joomla
  * @subpackage  Fabrik
- * @copyright   Copyright (C) 2005-2016  Media A-Team, Inc. - All rights reserved.\n * @author     Rob Clayburn
+ * @copyright   Copyright (C) 2005-2016  Media A-Team, Inc. - All rights reserved.
+ * @author     Rob Clayburn
  * @license     GNU/GPL http://www.gnu.org/copyleft/gpl.html
  */
 
@@ -28,6 +29,20 @@ class Com_FabrikInstallerScript
 	 * @var array
 	 */
 	protected $drivers = array('mysql_fab.php', 'mysqli_fab.php', 'pdomysql_fab.php');
+
+	/**
+	 * Documents >3.7
+	 *
+	 * @var array
+	 */
+	protected $documents38 = array('Partial', 'Pdf');
+
+	/**
+	 * Documents <= 3.7
+	 *
+	 * @var array
+	 */
+	protected $documents37 = array('fabrikfeed', 'partial', 'pdf');
 
 	/**
 	 * Run when the component is installed
@@ -226,23 +241,47 @@ class Com_FabrikInstallerScript
 	{
 		jimport('joomla.filesystem.folder');
 		jimport('joomla.filesystem.file');
-		$dest = JPATH_SITE . '/libraries/joomla/document/fabrikfeed';
 
-		if (JFolder::exists($dest))
+		// <= 3.7 documents formats
+		$dest = JPATH_ROOT . '/libraries/joomla/document';
+
+		foreach ($this->documents37 as $document)
 		{
-			if (!JFolder::delete($dest))
+			if (!empty($document) && JFolder::exists($dest . '/' . $document))
 			{
-				return;
+				JFolder::delete($dest . '/' . $document);
 			}
 		}
 
-		$dest = JPATH_SITE . '/libraries/joomla/database/database';
+		// database drivers, all versions
+		$dest = JPATH_ROOT . '/libraries/joomla/database/database';
 
 		foreach ($this->drivers as $driver)
 		{
-			if (JFile::exists($dest . '/' . $driver))
+			if (!empty($driver) && JFile::exists($dest . '/' . $driver))
 			{
 				JFile::delete($dest . '/' . $driver);
+			}
+		}
+
+		// >= 3.8 documents and renderers
+		$dest = JPATH_ROOT . '/libraries/src/Document';
+
+		foreach ($this->documents38 as $document)
+		{
+			if (!empty($document) && JFile::exists($dest . '/' . $document . 'Document.php'))
+			{
+				JFile::delete($dest . '/' . $document . 'Document.php');
+			}
+		}
+
+		$dest = JPATH_ROOT . '/libraries/src/Document/Renderer';
+
+		foreach ($this->documents38 as $document)
+		{
+			if (!empty($document) && JFolder::exists($dest . '/' . $document))
+			{
+				JFolder::delete($dest . '/' . $document);
 			}
 		}
